[runtime] Move ConfigSource classess
This commit moves ConfigSource (interface, general, env) to misc library.
